+++ NEWS.md
@@ -1,6 +1,7 @@
 # DedooseR 0.0.4
 * 2025-10-03: Added the plot_compare_saturation function so we can visualize the codes that meet criteria for thresholds set
 * 2025-10-03: Added the compare_saturation function so we can set saturation thresholds and see if codes meet them
+* 2025-10-03: Added the plot_saturation function so we can visualize set_saturation
 * 2025-10-03: Updated the set_saturation function so we can set by either code counts, transcript coverage or both
 
 # DedooseR 0.0.3